Refine this diagnostic test
